(RE-8186) Add Pkg::Util::Execution.capture3
This is a friendly wrapper around Open3.capture3. It maintains most of
the functionality of Pkg::Util::Execution.ex but the return matches that
of Open3.capture3 (stdout, stderr, retval). This is intended to allow
for more useful error messages in the event a command failed. Since
packaging is cross-loaded with enterprise-dist, making changes to the
ex() command is dangerous, but adding capture3 allows things to be
migrated over more gradually.