(RE-8186) Using pipes in Open3.capture3 with JRuby is special
After the initial work for this was merged in, failures became apparent
with razor packaging. Razor builds using JRuby 1.7.19. Whereas with ruby
2.1.8, we see
```
irb(main):004:0> Open3.capture3("echo 'hi' >/dev/null")
=> ["", "", #<Process::Status: pid 95032 exit 0>]
```
with JRuby 1.7.9 the same command results in
```
irb(main):005:0> Open3.capture3("echo 'hi' >/dev/null")
=> ["hi >/dev/null\n", "", #<Process::Status: pid=95178,exited(0)>]
```

This commit updates unnecessary piping in the code that checks
versioning and updates the method for curling to discard stdout when
curling quietly rather than dealing with pipes.

There are still pipes elsewhere in the code but this removes them from
enough places to allow razor to build successfully, and as that is the
only project that builds using jruby this should be sufficient.