(RE-6628) Update nightly_repo job with new pack_signed_repo tasks
Our automation is going to need to handle more that one set of platforms to
promote to PE. Instead of adding seperate jobs for seperate pipelines (i.e.
one for master, one for stable etc.) and continuing to carry around static
lists on top of information we already have, this commit adds the ability to
use the platform information in build_defaults.yaml to identify what platforms
promote.

This is done through a new pack_all_signed_repos_indivudually task, to replace
the pack_signed_repo task (that created the need for the list in the jenkins
job).

to faciliate this, the platform_repos data has been added to the available
data loaded in build_defaults.yaml, this data is in the following form:

platform_repos:
  - name: platform-version-arch
    repo_location: repos/platform/repo/location